Names refactor and added placeholder game card
