+++ src/modules/players/components/avatar.js
