+++ src/modules/players/components/list.js
