+++ src/modules/tournaments/components/list.js
