+++ src/modules/tournaments/views/index.js
@@ -1,3 +1,3 @@
-import TournamentList from '../components/tournament_list'
+import TournamentList from '../components/list'
 
 export default TournamentList
